refactor and split tests into separate classes
